# schema/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Infrahub",
        "version": "latest"
    },
    "paths": {
...
        "/api/transform/python/{transform_id}": {
            "get": {
                "summary": "Transform Python",
                "operationId": "transform_python_api_transform_python__transform_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    },
                    {
                        "APIKeyHeader": []
                    }
                ],
                "parameters": [
                    {
                        "name": "transform_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Transform Id"
                        }
                    },
                    {
                        "name": "branch",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Name of the branch to use for the query",
                            "title": "Branch"
                        },
                        "description": "Name of the branch to use for the query"
                    },
                    {
                        "name": "at",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Time to use for the query, in absolute or relative format",
                            "title": "At"
                        },
                        "description": "Time to use for the query, in absolute or relative format"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/transform/jinja2/{transform_id}": {
            "get": {
                "summary": "Transform Jinja2",
                "operationId": "transform_jinja2_api_transform_jinja2__transform_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    },
                    {
                        "APIKeyHeader": []
                    }
                ],
                "parameters": [
                    {
                        "name": "transform_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "ID or Name of the Jinja2 Transform to render",
                            "title": "Transform Id"
                        },
                        "description": "ID or Name of the Jinja2 Transform to render"
                    },
                    {
                        "name": "branch",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Name of the branch to use for the query",
                            "title": "Branch"
                        },
                        "description": "Name of the branch to use for the query"
                    },
                    {
                        "name": "at",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Time to use for the query, in absolute or relative format",
                            "title": "At"
                        },
                        "description": "Time to use for the query, in absolute or relative format"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            },
            "APIKeyHeader": {
                "type": "apiKey",
                "in": "header",
                "name": "X-INFRAHUB-KEY"
            }
        }
    }
}